+++ styleguide.config.js
@@ -68,7 +68,8 @@ module.exports = {
         'src/components/Input/index.js',
         'src/components/InputFile/index.js',
         'src/components/Radio/index.js',
-        'src/components/Select/index.js',
+        'src/components/BasicSelect/index.js',
+        'src/components/AdvancedSelect/index.js',
         'src/components/TextArea/index.js',
       ],
     }, {