+++ webpack/web.config.babel.js
@@ -1,6 +1,6 @@
 import merge from 'webpack-merge'
 
-// import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
 
 import webpackConfig from './base.config'
 
@@ -13,6 +13,6 @@ export default merge(webpackConfig, {
     extensions: ['.web.js', '.js']
   },
   plugins: [
-    // new BundleAnalyzerPlugin()
+    new BundleAnalyzerPlugin()
   ]
 })